chore: move homeDirectory into root.go
